Agregando buenas practicas, usando ResponseEntity, hicimos payload para los mensaje personalizados, y validamos cada metodo http si es necesario tanto en producto como categorias

+++ gestion-productos/src/main/java/com/proyecto/gestion/controlador/CategoriaControlador.java
@@ -2,9 +2,9 @@ package com.proyecto.gestion.controlador;
 
 import java.util.List;
 import java.util.stream.Collectors;
-
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,9 +14,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.proyecto.gestion.dto.CategoriaDTO;
-import com.proyecto.gestion.dto.ProductoDTO;
-import com.proyecto.gestion.entidad.Categoria;
+import com.proyecto.gestion.model.dto.CategoriaDTO;
+import com.proyecto.gestion.model.entidad.Categoria;
+import com.proyecto.gestion.model.payload.MensajeResponse;
 import com.proyecto.gestion.servicio.CategoriaServicio;
 
 import lombok.RequiredArgsConstructor;
@@ -31,41 +31,95 @@ public class CategoriaControlador {
 	
 	@GetMapping("categorias")
 	@ResponseStatus(HttpStatus.OK)
-	public List<CategoriaDTO> todasCategoria(){
+	public ResponseEntity<?> todasCategoria(){
 		
 		List<Categoria> categorias =  categoriaServicio.listarCategorias();
 		
-		return categorias.stream().map(categoria -> CategoriaDTO.builder()
+		List<CategoriaDTO> categoriasLista = categorias.stream().map(categoria -> CategoriaDTO.builder()
 				.id(categoria.getId())
 				.nombre_cat(categoria.getNombre_cat())
 				.build())
 				.collect(Collectors.toList());
 				
-				
-	
+			return new ResponseEntity<>(MensajeResponse.builder()
+					.mensaje("Consulta Exitosa")
+					.object(categoriasLista)
+					.build(), 
+					HttpStatus.OK);
 		}
 	
 	
 	
 	@PostMapping("categoria")
 	@ResponseStatus(HttpStatus.CREATED)
-	public Categoria agregarCategoria(@RequestBody CategoriaDTO categoriaDto) {
-		
-		Categoria categoria = categoriaServicio.agregarCategoria(categoriaDto);
+	public ResponseEntity<?> agregarCategoria(@RequestBody CategoriaDTO categoriaDto) {
 		
 		
-		return Categoria.builder()
-				.id(categoria.getId())
-				.nombre_cat(categoria.getNombre_cat())
-				.build();
+		Categoria categoria = null;
 		
+		
+		try {
+			
+			categoria = categoriaServicio.agregarCategoria(categoriaDto);
+			
+			Categoria nuevaCat = Categoria.builder()
+					.id(categoria.getId())
+					.nombre_cat(categoria.getNombre_cat())
+					.build();
+			
+			return new ResponseEntity<>(MensajeResponse.builder()
+					.mensaje("Categoria creada")
+					.object(nuevaCat)
+					.build(), 
+					HttpStatus.CREATED);
+			
+			
+		}catch(DataAccessException exDt) {
+		
+			return new ResponseEntity<>(MensajeResponse.builder()
+					.mensaje(exDt.getMessage())
+					.object(null)
+					.build(), 
+					HttpStatus.INTERNAL_SERVER_ERROR);
+			
+		}	
 	}
 	
 	@DeleteMapping("categoria/{id}")
 	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void eliminarCategoria(@PathVariable Integer id) {
+	public ResponseEntity<?> eliminarCategoria(@PathVariable Integer id) {
 		
-		categoriaServicio.eliminarCategoria(id);
+		try {
+			
+			if(categoriaServicio.existePorId(id)) {
+				
+				categoriaServicio.eliminarCategoria(id);
+				
+				return new ResponseEntity<>(MensajeResponse.builder()
+						.mensaje("Categoria eliminada")
+						.object(null)
+						.build(), 
+						HttpStatus.OK);
+				
+			}
+			
+			return new ResponseEntity<>(MensajeResponse.builder()
+					.mensaje("Categoria no existe")
+					.object(null)
+					.build(), 
+					HttpStatus.NOT_FOUND);
+			
+			
+		}catch(DataAccessException exDt) {
+			
+			
+			return new ResponseEntity<>(MensajeResponse.builder()
+					.mensaje(exDt.getMessage())
+					.object(null)
+					.build(), 
+					HttpStatus.INTERNAL_SERVER_ERROR);
+		
+		}
 	}
 	
 }
